Move NettyPipelinedConnection from http-netty to transport-netty-internal
Motivation
----------
The NettyPipelinedConnection has no knowledge or implementation details
which would tie it to the http package, so it can be moved to
transport-netty-internal where also the related NettyConnection
itself lives.

Modifications
-------------
Moved the NettyPipelinedConnection to the package, adjusted imports
and also moved the test case.

Result
------
The NettyPipelinedConnection can now also be used by other protocol
implementations and lives logically next to the NettyConnection
class itself.
