+++ servicetalk-transport-netty-internal/gradle/spotbugs/main-exclusions.xml
@@ -97,6 +97,11 @@
     <Method name="lambda$static$0"/>
     <Bug pattern="THROWS_METHOD_THROWS_CLAUSE_BASIC_EXCEPTION"/>
   </Match>
+  <!-- Parameters/state is intentional -->
+  <Match>
+    <Class name="io.servicetalk.transport.netty.internal.NettyPipelinedConnection"/>
+    <Bug pattern="EI_EXPOSE_REP2"/>
+  </Match>
 
   <!-- FIXME: 0.43 - Remove temporary suppression after we can remove deprecated constructors -->
   <Match>
